forgot one int64_t

+++ 2024/13.c
@@ -13,7 +13,7 @@
  * Get minimum runtime from timer output:
  *     m=999999;for((i=0;i<10000;++i));do t=$(./a.out|tail -n1|awk '{print $2}');((t<m))&&m=$t&&echo "$m ($i)";done
  * Minimum runtime measurements:
- *     Mac Mini 2020 (M1 3.2 GHz)                       :  5.29 µs
+ *     Mac Mini 2020 (M1 3.2 GHz)                       :  5.25 µs
  *     Raspberry Pi 5 (2.4 GHz)                         : 11.5  µs
  *     Macbook Air 2013 (Core i5 Haswell 4250U 1.3 GHz) : 26.8  µs
  *     Raspberry Pi 4 (1.8 GHz)                         : 28.6  µs
@@ -68,7 +68,7 @@ static int64_t tokens(const int part)
         printf("\n%3d: a(%"PRId64",%"PRId64") b(%"PRId64",%"PRId64") p(%"PRId64",%"PRId64")\n",
             i, c->a.x, c->a.y, c->b.x, c->b.y, c->p.x, c->p.y);
     #endif
-        const int d = c->a.x * c->b.y - c->a.y * c->b.x;  // divisor of inverted matrix
+        const int64_t d = c->a.x * c->b.y - c->a.y * c->b.x;  // divisor of inverted matrix
         lldiv_t n = lldiv(c->b.y * c->p.x - c->b.x * c->p.y, d);
         lldiv_t m = lldiv(c->a.x * c->p.y - c->a.y * c->p.x, d);
         // Must have no remainders, and quot<=100 for part 1.
